# backend/src/swagger.yaml
openapi: "3.0.3"
info:
  title: Plotra API
  description: 소설 창작 AI 보조 플랫폼 - REST API
  version: "1.0.0"
servers:
  - url: http://localhost:4000
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List projects
      operationId: listProjects
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
    post:
      tags: [Projects]
      summary: Create project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
                genre: { type: string }
                settings: { type: object }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /api/projects/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Projects]
      summary: Get project
      operationId: getProject
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Projects]
      summary: Update project
      operationId: updateProject
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                genre: { type: string }
                settings: { type: object }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Projects]
      summary: Delete project (soft)
      operationId: deleteProject
      responses:
        "200": { $ref: "#/components/responses/Deleted" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── Chapters ───
...
  /api/projects/{projectId}/plots:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Plots]
      summary: List plot items
      operationId: listPlotItems
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PlotItem" }
    post:
      tags: [Plots]
      summary: Create plot item
      operationId: createPlotItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, type]
              properties:
                title: { type: string }
                type:
                  type: string
                  enum: [synopsis, act, event, foreshadowing, conflict, ending_candidate, scene, other]
                content: { type: string }
                sortOrder: { type: integer }
                metadata: { type: object }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlotItem" }
  /api/projects/{projectId}/plots/{id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Plots]
      summary: Get plot item
      operationId: getPlotItem
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlotItem" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Plots]
      summary: Update plot item
      operationId: updatePlotItem
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                type:
                  type: string
                  enum: [synopsis, act, event, foreshadowing, conflict, ending_candidate, scene, other]
                content: { type: string }
                sortOrder: { type: integer }
                metadata: { type: object }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlotItem" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Plots]
      summary: Delete plot item (soft)
      operationId: deletePlotItem
      responses:
        "200": { $ref: "#/components/responses/Deleted" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── Plot ↔ Chapters ───
  /api/projects/{projectId}/plots/{id}/chapters:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Plot Chapters]
      summary: List linked chapters
      operationId: listPlotItemChapters
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PlotItemChapter" }
    post:
      tags: [Plot Chapters]
      summary: Link chapter to plot item
      operationId: createPlotItemChapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chapterId]
              properties:
                chapterId: { type: string, format: uuid }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlotItemChapter" }
  /api/projects/{projectId}/plots/{id}/chapters/{linkId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/Id"
      - $ref: "#/components/parameters/LinkId"
    delete:
      tags: [Plot Chapters]
      summary: Unlink chapter from plot item
      operationId: deletePlotItemChapter
      responses:
        "200": { $ref: "#/components/responses/Deleted" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── Plot ↔ Characters ───
  /api/projects/{projectId}/plots/{id}/characters:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Plot Characters]
      summary: List linked characters
      operationId: listPlotItemCharacters
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PlotItemCharacter" }
    post:
      tags: [Plot Characters]
      summary: Link character to plot item
      operationId: createPlotItemCharacter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [characterId]
              properties:
                characterId: { type: string, format: uuid }
                role: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlotItemCharacter" }
  /api/projects/{projectId}/plots/{id}/characters/{linkId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/Id"
      - $ref: "#/components/parameters/LinkId"
    delete:
      tags: [Plot Characters]
      summary: Unlink character from plot item
      operationId: deletePlotItemCharacter
      responses:
        "200": { $ref: "#/components/responses/Deleted" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── Notes ───
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "Encrypted JWT token"
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    LinkId:
      name: linkId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    Deleted:
      description: Deleted
      content:
        application/json:
          schema:
            type: object
            properties:
              deleted: { type: boolean }
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        title: { type: string }
        description: { type: string, nullable: true }
        genre: { type: string, nullable: true }
        settings: { type: object }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
...
    PlotItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        title: { type: string }
        type:
          type: string
          enum: [synopsis, act, event, foreshadowing, conflict, ending_candidate, scene, other]
        content: { type: string }
        sortOrder: { type: integer }
        metadata: { type: object }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
    PlotItemChapter:
      type: object
      properties:
        id: { type: string, format: uuid }
        plotItemId: { type: string, format: uuid }
        chapterId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
    PlotItemCharacter:
      type: object
      properties:
        id: { type: string, format: uuid }
        plotItemId: { type: string, format: uuid }
        characterId: { type: string, format: uuid }
        role: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...